finalise implementation, add documentation
